src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "PicASpot backend"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.0",
    # asyncpg is not compiled on Windows (for now, waiting for wheels).
    # We compile it on Docker Linux containers instead
    "asyncpg>=0.30.0; sys_platform != 'win32'",
    "asgiref>=3.10.0",
    "bcrypt>=5.0.0",
    "celery>=5.5.3",
    "celery-types==0.23.0",
    "cryptography>=46.0.3",
    "fastapi-mail>=1.5.7",
    "fastapi[standard]>=0.119.0",
    "flower>=2.0.1",
    "geoalchemy2[shapely]>=0.18.0",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "redis>=7.0.0",
    "sqlalchemy>=2.0.44",
]


[project.optional-dependencies]
test = [
    "faker>=37.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[tool.isort]
profile = "black"
skip_gitignore = true
known_first_party = ["app", "tests"]

[tool.ruff]
target-version = "py314"
indent-width = 4

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # complexity
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
